Update XSContentModel.java
Fixed a bug whereby the content model for minOccurs=1, maxOccurs=1000 was not being emitted properly as part of the content model definition. Also now emitting attributes at part of the element definition in the content model. 

Line 84 -> Changed the code from: 

            else if(particle.getMinOccurs()!=1 && particle.getMaxOccurs()!=-1)
                buff.append("[").append(particle.getMinOccurs()).append(",").append(particle.getMaxOccurs()).append("]");

to 

            else if(particle.getMinOccurs()!=particle.getMaxOccurs())
                buff.append("[").append(particle.getMinOccurs()).append(",").append(particle.getMaxOccurs()).append("]");

Line 116: implemented postProcess method, to allow for the generation of attributes as part of the content model: 

For example, ServicePackage element below defined 3 attributes:
 * ID0 - mandatory
 * ID1 - optional
 * ID2 - mandatory. 

((ns:ServicePackage[0,1000] @ID0 @ID1? @ID2 , ns:MSISDNList[1,1000]))